pyproject.toml
[tool.poetry]
name = "rchi"
version = "0.1.0"
description = ""
authors = ["rohan chakraborty <rohanchakraborty123@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
torch = "^2.5.1"
torchvision = "^0.20.1"
numpy = "^2.1.3"
lightning = "^2.4.0"
datasets = "^3.1.0"
tqdm = "^4.67.1"
matplotlib = "^3.9.3"
huggingface-hub = "^0.26.5"
rich = "^13.9.4"
loguru = "^0.7.3"
python-dotenv = "^1.0.1"


[tool.poetry.group.dev.dependencies]
ruff = "^0.8.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
